move 2 tests to test_cli.py
