+++ tests/test_cli.py
@@ -16,6 +16,7 @@ from rich.progress import (
 )
 
 from wtfis.clients.abuseipdb import AbuseIpDbClient
+from wtfis.clients.base import requests
 from wtfis.clients.greynoise import GreynoiseClient
 from wtfis.clients.ip2whois import Ip2WhoisClient
 from wtfis.clients.ipwhois import IpWhoisClient
@@ -23,12 +24,14 @@ from wtfis.clients.shodan import ShodanClient
 from wtfis.clients.urlhaus import UrlHausClient
 from wtfis.clients.virustotal import VTClient
 from wtfis.config import Config, parse_args, parse_env
-from wtfis.exceptions import WtfisException
+from wtfis.exceptions import HandlerException, WtfisException
 from wtfis.handlers.domain import DomainHandler
 from wtfis.handlers.ip import IpAddressHandler
-from wtfis.main import generate_entity_handler, generate_view, main
+from wtfis.main import fetch_data, generate_entity_handler, generate_view, main
 from wtfis.models.virustotal import Domain, IpAddress
 from wtfis.ui.view import DomainView, IpAddressView
+
+# pylint: disable=protected-access,redefined-outer-name
 
 POSSIBLE_ENV_VARS = [
     "VT_API_KEY",
@@ -774,3 +777,70 @@ class TestMain:
         m_view.assert_called_once_with(m_conf(), m_console(), m_handler())
         m_view().print.assert_called_once_with(one_column=False)
         unset_env_vars()
+
+
+class TestFetchData:
+    """
+    Test main.fetch_data()
+    """
+
+    @patch.object(requests.Session, "get")
+    def test_vt_http_error(self, mock_requests_get, domain_handler, capsys):
+        """
+        Test a requests HTTPError from the VT client. This also tests the
+        common_exception_handler decorator.
+        """
+        handler = domain_handler()
+        mock_resp = requests.models.Response()
+
+        mock_resp.status_code = 401
+        mock_requests_get.return_value = mock_resp
+
+        # Thorough test of first _fetch_* method
+        with pytest.raises(SystemExit) as e:
+            fetch_data(MagicMock(), handler)
+
+        capture = capsys.readouterr()
+
+        assert (
+            capture.err == "Error fetching data: 401 Client Error: None for url: None\n"
+        )
+        assert e.type is SystemExit  # ruff E721
+        assert e.value.code == 1
+
+        # Extra: just make sure program exits correctly
+        with pytest.raises(HandlerException) as e:
+            handler._fetch_vt_resolutions()
+
+    @patch.object(requests.Session, "get")
+    def test_vt_validation_error(self, mock_requests_get, domain_handler, capsys):
+        """
+        Test main.fetch_data().
+        Test a pydantic data model ValidationError from the VT client. This also tests
+        the common_exception_handler decorator.
+        """
+        handler = domain_handler()
+        mock_resp = requests.models.Response()
+
+        with patch.object(mock_resp, "json") as mock_resp_json:
+            mock_resp.status_code = 200
+            mock_resp_json.return_value = {"intentionally": "wrong data"}
+            mock_requests_get.return_value = mock_resp
+
+            # Thorough test of first _fetch_* method
+            with pytest.raises(SystemExit) as e:
+                fetch_data(MagicMock(), handler)
+
+            capture = capsys.readouterr()
+
+            assert capture.err.startswith(
+                "Data model validation error: 1 validation error for Domain\ndata\n"
+                "  Field required [type=missing, input_value={'intentionally': "
+                "'wrong data'}, input_type=dict]\n"
+            )
+            assert e.type is SystemExit
+            assert e.value.code == 1
+
+            # Extra: just make sure program exits correctly
+            with pytest.raises(HandlerException) as e:
+                handler._fetch_vt_resolutions()
